requirements.txt
asgiref==3.8.1
backports-datetime-fromisoformat==2.0.3
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
environs==14.2.0
inflection==0.5.1
marshmallow==4.0.0
packaging==25.0
PyJWT==2.9.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
